+++ packages/vseed/src/pipeline/advanced/chart/pipes/encoding/rose.ts
@@ -66,9 +66,7 @@ const generateMeasureEncoding = (measures: Measures, encoding: Encoding) => {
 }
 
 const generateDimensionEncoding = (dimensions: Dimensions, encoding: Encoding) => {
-  encoding.angle = unique(
-    dimensions.filter((item) => item.encoding === 'angle' || !item.encoding).map((item) => item.id),
-  )
+  encoding.angle = unique(dimensions.filter((item) => item.encoding === 'angle').map((item) => item.id))
   encoding.color = unique(dimensions.filter((item) => item.encoding === 'color').map((item) => item.id))
   encoding.detail = unique(dimensions.filter((item) => item.encoding === 'detail').map((item) => item.id))