+++ packages/vseed/src/pipeline/spec/chart/pipes/markStyle/funnelStyle.ts
@@ -1,0 +1,25 @@
+import type { IFunnelChartSpec } from '@visactor/vchart'
+import type { SpecPipe } from 'src/types'
+
+export const funnelStyle: SpecPipe = (spec) => {
+  const result = {
+    ...spec,
+    funnel: {
+      style: {},
+    },
+  } as IFunnelChartSpec
+
+  return {
+    ...result,
+    funnel: {
+      style: {
+        cornerRadius: 4,
+      },
+      state: {
+        hover: {
+          fillOpacity: 0.6,
+        },
+      },
+    },
+  }
+}