+++ packages/vseed/src/pipeline/spec/chart/pipes/markStyle/roseStyle.ts
@@ -1,0 +1,31 @@
+import type { IRoseChartSpec } from '@visactor/vchart'
+import type { SpecPipe } from 'src/types'
+
+export const roseStyle: SpecPipe = (spec, context) => {
+  const { advancedVSeed } = context
+  const { dataset } = advancedVSeed
+  const showStroke = dataset.length <= 30
+
+  const result = {
+    ...spec,
+    rose: {
+      style: {},
+    },
+  } as Required<IRoseChartSpec>
+
+  return {
+    ...result,
+    rose: {
+      style: {
+        stroke: '#ffffff',
+        lineWidth: showStroke ? 1 : 0,
+      },
+      state: {
+        hover: {
+          lineWidth: 1,
+          fillOpacity: 0.6,
+        },
+      },
+    },
+  }
+}